Refactor query builder
